LayoutTests/media/track/audio/audio-track-mkv-vorbis-addtrack.html

<!DOCTYPE html>
<html>
    <head>
        <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />

        <script src=../../media-file.js></script>
        <script src=../../video-test.js></script>
        <script src=../../in-band-tracks.js></script>
    </head>
    <body onload="testAddTrack('../../content/two-audio-and-video-tracks.mkv', 'audio')">
        <video controls></video>
        <p>Testing audio track "addtrack" events.</p>
    </body>
</html>
