LayoutTests/media/track/audio/audio-track-mkv-vorbis-addtrack-expected.txt


Testing audio track "addtrack" events.

EVENT(addtrack)
EXPECTED (video.audioTracks contains 'event.track') OK

EVENT(addtrack)
EXPECTED (video.audioTracks contains 'event.track') OK

EVENT(canplaythrough)
END OF TEST
